Refactor tests to reduce duplicate code while maintaining coverage, add jacoco
