+++ crypto/build.gradle
@@ -1,6 +1,7 @@
 plugins {
     id "java-library"
     id "io.freefair.lombok" version "6.0.0-m2"
+    id 'jacoco'
 }
 
 sourceCompatibility = JavaVersion.VERSION_11
@@ -24,6 +25,7 @@ dependencies {
 
 test {
     useJUnitPlatform()
+    finalizedBy jacocoTestReport // report is always generated after tests run
 }
 
 java {
@@ -33,7 +35,7 @@ java {
 
 test {
     useJUnitPlatform()
-    exclude '**/benchmark/**'
+    exclude '**/benchmark/run/**'
 }
 
 task benchmark(type: Test) {
@@ -42,5 +44,5 @@ task benchmark(type: Test) {
         outputs.upToDateWhen { false }
         showStandardStreams = true
     }
-    include '**/benchmark/**'
+    include '**/benchmark/run/**'
 }
