+++ crypto/src/test/java/org/cryptimeleon/incentive/crypto/benchmark/BenchmarkTest.java
@@ -1,5 +1,6 @@
-package org.cryptimeleon.incentive.crypto;
+package org.cryptimeleon.incentive.crypto.benchmark;
 
+import org.cryptimeleon.incentive.crypto.Setup;
 import org.cryptimeleon.incentive.crypto.benchmark.Benchmark;
 import org.cryptimeleon.incentive.crypto.benchmark.BenchmarkConfig;
 import org.cryptimeleon.incentive.crypto.benchmark.BenchmarkResult;
