+++ crypto/src/test/java/org/cryptimeleon/incentive/crypto/benchmark/run/BenchmarkTask.java
@@ -1,6 +1,9 @@
-package org.cryptimeleon.incentive.crypto.benchmark;
+package org.cryptimeleon.incentive.crypto.benchmark.run;
 
 import org.cryptimeleon.incentive.crypto.Setup;
+import org.cryptimeleon.incentive.crypto.benchmark.Benchmark;
+import org.cryptimeleon.incentive.crypto.benchmark.BenchmarkConfig;
+import org.cryptimeleon.incentive.crypto.benchmark.BenchmarkResult;
 import org.junit.jupiter.api.Test;
 
 import java.util.logging.Logger;
