+++ crypto/src/test/java/org/cryptimeleon/incentive/crypto/model/PublicParameterTest.java
@@ -1,5 +1,7 @@
-package org.cryptimeleon.incentive.crypto;
+package org.cryptimeleon.incentive.crypto.model;
 
+import org.cryptimeleon.incentive.crypto.IncentiveSystem;
+import org.cryptimeleon.incentive.crypto.Setup;
 import org.cryptimeleon.incentive.crypto.model.IncentivePublicParameters;
 import org.junit.jupiter.api.Test;
